requirements.txt
requests
pycryptodome
numpy
urllib